+++ ProjectTRM/Objects/Character/Player/Tank/P_Tank.h
@@ -80,7 +80,7 @@ public:
 	/// HPŠÇ—ˆ—
 	/// </summary>
 	/// <param name="hit_object">ƒ_ƒ[ƒW</param>
-	virtual void HPControl(int Damage) override;
+	virtual void HPControl(float Damage) override;
 
 private:
 	/// <summary>
